python-flask-restapi/pyproject.toml
[tool.poetry]
name = "python-flask-restapi"
version = "0.1.0"
description = ""
authors = ["samuel valverde <savalverde@estudiantec.cr>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11.8"
Flask = "^3.0.2"
python-dotenv = "^1.0.1"
psycopg2-binary = "^2.9.9"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
